+++ tests/Unit/DependencyInjection/TsfGatekeeperAiExtensionTest.php
@@ -10,6 +10,9 @@ use Tsf\GatekeeperAiBundle\Command\ValidateCommand;
 use Tsf\GatekeeperAiBundle\DependencyInjection\TsfGatekeeperAiExtension;
 use Tsf\GatekeeperAiBundle\Installer;
 use Tsf\GatekeeperAiBundle\Service\Config\Settings;
+use Tsf\GatekeeperAiBundle\Service\Provider\Anthropic\AnthropicProvider;
+use Tsf\GatekeeperAiBundle\Service\Provider\EnrichmentProviderInterface;
+use Tsf\GatekeeperAiBundle\Service\Provider\FakeProvider;
 use Tsf\GatekeeperAiBundle\Service\ProposalStore;
 
 final class TsfGatekeeperAiExtensionTest extends Unit
@@ -28,5 +31,14 @@ final class TsfGatekeeperAiExtensionTest extends Unit
         $config = $container->getDefinition(Settings::class)->getArgument('$config');
         self::assertSame(['title'], $config['classes']['Product']['enrich']);
         self::assertSame($config, $container->getParameter('tsf_gatekeeper_ai.config'));
+        self::assertSame(AnthropicProvider::class, (string) $container->getAlias(EnrichmentProviderInterface::class));
+    }
+
+    public function testTheFakeProviderCanBeSelected(): void
+    {
+        $container = new ContainerBuilder();
+        (new TsfGatekeeperAiExtension())->load([['provider' => 'fake']], $container);
+
+        self::assertSame(FakeProvider::class, (string) $container->getAlias(EnrichmentProviderInterface::class));
     }
 }